Refine README.md and refactor code structure for "machine translation with external memory" model.
